+++ src/qq_mem/src/utils.h
@@ -180,10 +180,10 @@ std::string format_with_commas(T value)
 
 
 std::string str_qq_search_reply(const qq::SearchReply &reply);
-int varint_expand_and_encode(uint32_t value, std::string *buf, const int offset);
-int varint_encode(uint32_t value, std::string *buf, int offset);
-int varint_decode(const std::string &buf, int offset, uint32_t *value);
-int varint_decode_chars(const char *buf, const int offset, uint32_t *value);
+int varint_expand_and_encode(uint32_t value, std::string *buf, const off_t offset);
+int varint_encode(uint32_t value, std::string *buf, off_t offset);
+int varint_decode(const std::string &buf, off_t offset, uint32_t *value);
+int varint_decode_chars(const char *buf, const off_t offset, uint32_t *value);
 
 void MapFile(std::string path, char **ret_addr, int *ret_fd, size_t *ret_file_length);
 void UnmapFile(char *addr, int fd, size_t file_length);
